Release of "skeleton" sensor_fusion application
This is an incomplete release of sensor_fusion. It provides a simple interface with a ROS communicating node.
This sensor fusion node will subscribe to drivers it is intererested and BSM messages. Right now it just assumes
there is only one driver providing data and simply republished the messages. BSM data has not been incorporated into
the connected_vehicles per the architecture design, this will be updated as soon as a BSM message format is received