Refactor guidance into guidance and guidance_plugin_api
guidance_plugin_api is now to be the ROS agnostic JAR file we distribute to for plugin development
It has a facade API over top of the whole PubSub api as well as interfaces for each of the core
guidance components. These interfaces will need to be kept in step with the actual classes as
development progresses and will also serve as our dependency injection seams.