Fix state check in guidance state callback
Guidance state updates were being checked against == INACTIVE which may miss some cases in which the trajectory should not be consumed. This has been fixed to check against != ENGAGED. A redundant callback (unused) that contained this logic has been removed and the correct callback has been updated with this logic.