Fix/proj tim (#1322)
Description
This PR updates the traffic incident parser to use the same TCM projection interpretation as carma-cloud for the RWM use case.
I also noticed an exception was thrown when the parsing failed so I added logic to protect against that case.

Related Issue
This PR resolves #1326

Motivation and Context
Both TIM and RWM can function on the same build

How Has This Been Tested?
The ford fusion was used to demonstrate both use cases successfully at summit point using this branch. This was including both lane changes and the speed reduction.