Feature/refactor stop and wait (#1353)
updates the Stop and Wait plugin to use a constant acceleration for planning stopping maneuvers instead of a constant jerk. This makes the calculations cleaner and the expected behavior easier to think about. To accomplish this change a number of modifications were needed

RouteFollowingPlugin updated to plan a Stop And Wait maneuver at the end of a route based on the vehicle longitudinal acceleration limit with a multiplier to keep the planned decel under the limit.
Stop And Wait plugin updated to use constant acceleration
Pure Pursuit Jerk Wrapper totally removed
Extra logic added to inlane-cruising to prevent overlapping points on lanelet transitions
negative speed guard added to pure_pursuit_wrapper
bug in world model was resolved such that route length is properly set base don the end point
update other packages to use the modified maneuver message