Feature/traffic signal (#1355)
The DSRC spat message is picked up by the WM interface and the map is updated with the traffic light data with real time update. The map should be updated at a higher frequency to accurately track the traffic light. The map is updated and then send to the guidance. When the guidance sees the stop sign. The stop and wait manure is invoked with the help of a strategic plugin. When the light turns green, the intersection transit manure is invoked. Note: we already have the lane keeping straight which just needs to be replaced with intersection transit. At the exit, the guidance will just switch to lane following or any plugin based on the situation.

Co-authored-by: Michael McConnell <Michael.McConnell-2@leidos.com>