Feature/add system controller (#1493)
This PR adds two packages to carma-platform the first is system_controller which is a lifecycle controller for the overall carma-platform system. Specifically it manages the subsystem controllers which are defined in the subsystem_controllers package. These controllers take a list of required nodes as well as a namespace and monitor the system alert topic for failures in these required components and or any node in their namespace. Based on the needs of the specific subsystem they can forward that failure to the overall system_controller which will shutdown carma if an entire subsystem fails. At the moment the decision making logic is relatively simple, but this structure allows us to add far more complex logic to be added to the relevant subsystem as needed.