Add localization_manager fault when GNSS data timesout and in a GNSS only state. (#1542)
This PR updates the localization_manager to have a timeout for the GNSS data when in the DEGRADED_NO_LIDAR_FIX state. If this occurs then an exception is thrown. This exception should cause a FATAL system_alert to be sent to the health_monitor and result in a system shutdown.

Related Issue
Fixes #1543