Feature/ros2 object detection (#1558)
This PR updates the CARMA Platform such that the docker image builds both the ROS1 and ROS2 version of the system.
In addition, the object_detection_tracking package was converted to ROS2 and a new ROS2 environment.launch.py and carma_src.launch.py file was created to launch this ROS2 object perception stack which leverages Autoware.Auto components.

This PR also contains previously reviewed code changes to add the template_package to the repo in this PR: #1555

This PR looks intimidating (128 files changed), but it can actually be approached in a handful of straight forward steps.

Everything under template_package has already been reviewed and there is no need to review further.
In the object_detection_tracking folder however the focus should be on the refactor rather than logic as this has not changed. So CMake, package, lifecycle behavior should be the focus.
The environment.launch.py and carma_src.launch.py files serve as new ROS2 launch files in need of review
Changes to the carma package are related to point 3. The new launch files were added and the old ROS1 nodes are no longer needed were removed.
A host of ROS1 changes were needed to message fields that had to be modified to support ROS1 bridge in this PR: https://github.com/usdot-fhwa-stol/carma-msgs There are no logic changes related to this only variable name changes. Compilation should be a pretty good check of accuracy for this but a quick review is always a good idea.
The guidance_command_repeater package was removed as it is left over from CARMA2 and remains totally unused.

* Initial implementation of Autoware.Auto object perception launch file.

* Remove unused deps from object_detection_tracking

* cleanup deps in object_detection_tracking

* Remove invalid include object_detection_tracking

* WIP most of object_detection_tracking converted

* Fix typo in build-image.sh for foxy-develop tag

* object_detection_tracking builds but itsn't installing cleanly

* Functional object_detection_tracking

* WIP launch file

* Functional environment.py launch file for object perception and tracking

* Create initial carma_src ros2 launch file

* Properly install carma package in ros2

* WIP top level launch file and updated install script

* WIP controllable logging

* wip

* Remove unused guidance_command_repeater

* WIP

* Buildable mobilitypath_publisher

* build fixes

* Full docker build

* cleanup launch scripts and add log level control

* Move get_log_level to scripts and improve error handling

* Feature/template package (#1555)

* template_package and backup

* Functional template_package

* Add example unit test

* Cleanup readme

* add some helpful comments

* Add missing license

* Add explicit runtime dependencies

* Return all deps to foxy develop

* Fully buildable hybrid carma platform docker image

* Address PR comments

Co-authored-by: afrye-leidos <alexander.frye@leidos.com>