Feature/plugin base library (#1782)
This PR introduces an implementation of the plugin base library for outlined in this design
https://usdot-carma.atlassian.net/wiki/spaces/CRMPLT/pages/2182545409/Detailed+Design+-+Plugin+Library

I've gone ahead and implemented it to facilitate the creation of a plugin controller in ROS2 to replace the health_monitor.

The core change (or really addition) to the design is the integration of the lifecycle state machine for plugins. Previously plugins had one core state which was ACTIVE. Now they follow the standard ROS2 lifecycle where they are active when the when their lifecycle state is active. To decouple the plugin states from the underlying state machine implementation the template pattern is used to hide the existing lifecycle callbacks. Plugins also have a concept of Availability which has not been fully defined in CARMA3. For this reason, the get_availability() function is provided but availability is not integrated into the full state machine. If we wish to implement it, that could be accomplished by adding it as an additional state on top of active.

To make usage of the plugin base class easier the template_package has also been updated to allow for the user to specify a plugin type "strategic", "tactical", "control" and generate the appropriate extending classes. If the type is not provided, the package will default to a standard node as before. There is a fair bit of template code duplication in the template_package now which I spent a good deal of time trying to avoid but ultimately decided it was the cleanest solution.

Related Issue
#1801