Feature/platooning strategic ros2 (#1894)
Description
This PR ports the platooning strategic ihp plugin to ROS2 as well as fixes a number of bugs associated with recent ports.
Changes:

platoon_strategic_ihp ported to ROS2
platoon_strategic removed as it has been fully replaced by platoon_strategic_ihp
Updates some platooning unit tests to compile and pass. They had been very out of date during ihp2 testing. There are still 7 failing tests which need to be updated by someone with more knowledge of the state machine. But these tests are also failing in ROS1.
ui.launch.py corrected to use the launch file from rosbrige_websocket. It was previously trying to load a python node as a component which is impossible.
Removes the UI parameters file which didn't have a clear path to loading and has never been modified in carma2 or 3 so defaults should be plenty sufficient.
Corrects a bug in the stop_and_wait_plugin where the wrong library was specified as the component making the class unloadable
Corrects a bug in route_following_plugin where a decimal parameter was marked as an integer
This PR should be merged simultaneously with
usdot-fhwa-stol/carma-config#207
and
usdot-fhwa-stol/carma-web-ui#148

Related Issue
Resolves #1895
Resolves #1896
Resolves #1897
Resolves #1898
Resolves #1899