Feature/uc3 edge cases (#1991)
Description
In LCI Plugin, given an ET, the vehicle will first find a smoothed trajectory for itself (trajectory cases 1-8). Once the trajectory case is initially determined, the logic has some checks to decide its action and desired speed. One of the cases is called HandleFailureCase() which will be triggered if the vehicle was following a trajectory case previously and suddenly its ET becomes out of reach when the vehicle is close to the intersection and its distance to the stop bar is less than the safe stopping speed. In this case, since the vehicle is not able to stop at the stop bar with the max comfortable deceleration rate, HandleFailureCase() tells the vehicle to try to accelerate\decelerate to enter the intersection box with a speed close to the departure speed.

There are two main scenarios that might trigger HandleFailureCase().

The vehicle ET is at the beginning of the green and the vehicle gets closer to the intersection that it should be based on its smoothed trajectory. In this case, the vehicle must slow down to allow the signal to change from red to green before entering the intersection box.
The vehicle ET is at the end of the green and as the vehicle gets closer to the intersection box, due to inaccuracy in trajectory control, the vehicle receives an ET that is pushed back to the beginning of the next green.

Related Issue
#1985

Motivation and Context
TSMO UC3 integration testing

How Has This Been Tested?
Integration testing in progress

Co-authored-by: Misheel Bayartsengel <misheel.bayart@gmail.com>