#2137 Minor improvements to CMake configuration
The PROJ4 CMake module has now been vendored into the
carma_cooperative_perception package in case it is not already installed
on the host system. The libproj-dev apt package does not provide it, so
users normally have to download it elsewhere.

Other changes include adding an option to export a compilation database
and creating a COLCON_IGNORE file in the CMake project's binary
directory.