#2140 Add linting configuration
The carma_cooperative_perception package now lints the code when testing
is enabled. This commit changes a lot of the existing code to satisfy
the linting requirements. Most of the changes include license header
relocation to the very top of the file, include guard name changes
according to the ROS style guide, and removing namespace
using-declarations.