#2140 Rename SDSM to detection list file
The previous name had `node` appended to it, which conflicts with the
naming convention of having node executables append `node` to their
filename. I recommend we use the naming convention `_component` for
nodes that are going to be implemented as ROS components. The proper
node executable source file can then have a `_node` suffix.