[cooperative perception] Add multiple object tracker node (#2165)
* #2164 Add skeleton for multiple object tracker

The multple object tracker node will ingest detections, execute the
tracking pipeline, then publish the resulting tracks. This commit
covers adds the node skeleton with the implementaiton coming in future
commits.

* #2164 Update function signatures

* #2164 Add functions to create Detections

These functions create CTRV and CTRA detections from ROS 2 messages.

* #2164 Move detection storage into function

The detection storage logic has been moved from the subscription
callback to a separate function because there is some processing that
should happen before the detections are possibly added to the internal
detection lis.

* #2164 Add Track to ROS 2 message function

This function creates ROS2 messages from the tracking pipeline's outputs

* #2164 Add initial multiple object tracker

Add initial multiple object tracker node implementation for performing
the pipeline execution.

* #2164 Add check for existing detection in list

If the detection is already in the list to be processed, the detection's
data is updated instead of being appended to the list.

* #2164 Add dynamic execution frequency

The executing frequency for the tracking pipeline can now be changed
while the node is running.

* #2164 Change track manager type

The multiple object tracker node now uses a simpler track manager that
combines the management policy with the manager itself. This is a change
in the underlying multiple object tracking library. We only needed to
adhere to the new type.

* #2164 Update parameter setting logging

This is really just a cosmetic change.

* #2164 Resolve linting errors

* #2164 Fix bug in track manager instance

The removal threshold was equal to the promotion threshold (and above
the initial occurrance count), so tracks were removed as soon as they
were added to the track manager.

* #2164 Change pipeline logic to always publish

Even if the track list is empty, we should always publish a message to
indicate that the tracker is functioning properly. Additionally, an
empty track list means there are no tracks perceived in the environment.

* #2164 Add clustering to tracking pipeline

We need clustering to choose which detection out of the group will get
turned into a track. Otherwise, we will have duplicate tracks.

* #2164 Add component tests

The component tests use static object scenarios.

* #2164 Fix covariance creation bug

When creating detections from ROS messages, the factory functions did
not properly zero-out the covariance matrices, which resulted in
numerical errors. The covariance matrices should be initialized to
the data from the message, but that is going to be more involved.

* #2164 Add Node documentation

* #2164 Fix typos

* #2164 Add multiple object tracking parameters

The additional parameters allow users to modify the tracker's
promotion and removal thresholds. An example params file has been added
along with a basic launch file.

* #2164 Address peer review about managed pubs/subs

* #2164 Add FetchContent for MOT library

The multiple_object_tracking library was not being installed as a
dependency, which is most likely what is causing the CI build to fail.

* #2164 Update multiple_object_tracking dependency

The carma_cooperative_perception library was previously installing the
library as a source dependency, but now colcon will build it for us.

* #2164 Add MOT library to CARMA checkout script

The multiple_object_tracking library is not available through a
package manager, so it needs to be downloaded from source.

* #2164 Update multiple_object_tracker dependency

The checkout script now downloads from the develop branch. The commented
out FetchContent configurations for carma_cooperative_perception have
been removed.