Add try/catch block around client creation (#2194)
* #2193 Add try/catch block around client creation

The try...catch block should catch any RCL-related issues when the
CapabilitiesInterface instance tries to create a ROS 2 service client
but fails.

* #2193 Update try catch to be more specific

The previous try...catch was catching the base RCL error, which is too
vague. It has been narrowed to RCLError since that is what we're
expecting to possibly get thrown.

* #2193 Move code inside try...catch

The previous implementation had compilation errors because some
variables were out of scope.

* #2193 Clean up CapabilitiesInterface function

This change cleans up the multiplex_service_call_for_capability
function. It does not change any of the behavior (besides the try...
catch).

* #2193 Address compiler warnings

These are minor things that don't affect behavior.