[cooperative perception] Move motion model mapping default values to `MotionModelMapping` class (#2203)
* #2202 Move default mapping values to class

The default motion model mapping values were specified in the ROS
parameter declarations, but this meant the MotionModelMapper instance
never got them. Its values wouldn't be updated unless a user changed
them. Now, the default values are stored in the class instance, and the
ROS parameter declarations get their default values from the instance.

* #2202 Update parameter declaration

The parameter declarations for the ExternalObjectListToDetectionListNode
got moved to after the on_set_parameter_callback assignment because
the callback was previously not being called when the parameters were
initially declared. This prevented the motion_model_mapping_ from
being updated with the user-overridden mappings.

* #2202 Change default motion model mapping

The previous motion model mapping used CV model for pedestrians and
and unknowns, but the CV model is unsupported. This changes the default
to CTRV so we don't default to unsupported motion models.