[cooperative perception] Add covariance calculations to SDSM conversions (#2293)
* CDAR-729 Set covariances from confidences

The SDSM to detection conversions now set the detections' state
covariances from the incoming SDSMs' confidences.

* CDAR-729 Add try...catch around conversion

There are a lot of optional fields in SDSMs that are required for our
purposes. Attempting to access one that isn't set will throw a
bad optional access error, so we wrap the conversion call and check for
that error.

* CDAR-729 Fix covariance conversions

Add checks for bad optional access. If an exception gets thrown,
propagate that back up to caller. Also fix the indices for the
covariance matrices.

* CDAR-729 Update comments