[cooperative perception] Compensate for CDASim time lag (#2304)
* CDAR-757 Add CDASim time source subscription

When simulating CARMA, there is a discrepancy between CARLA's time and
CDASim's time. The CP stack needs to compensate for this when
converting SDSM timestamps. The long term solution is to fix this at the
CDASim level, but this is what we're doing for now.

* CDAR-757 Fix detection header update bug

In the header update loop, detection was not declared as a reference
when it should have been.

* CDAR-757 Add topic remapping rule

We need to remap the input/cdasim_clock topic to /sim_clock so we
actually get the time.